LayoutTests/requestidlecallback/requestidlecallback-deadline-expected.txt

This tests that the deadline can be reached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS idleDeadline.timeRemaining() is 0
PASS idleDeadline.didTimeout is true
PASS didRunIdleCallback is true
PASS successfullyParsed is true

TEST COMPLETE
